chore: move `cssClasses.ts` to `@volar/vue-language-service`
